This looks great everyone!

+++ src/components/LessonCard.jsx
@@ -25,21 +25,25 @@ import { useTheme } from './ThemeContext';
 function LessonCard({ lesson, onDelete }) {
   const navigate = useNavigate();
   const toast = useToast();
-  const [selectedColor, setSelectedColor] = useState(lesson.tag || 'white');
+  const [selectedColor, setSelectedColor] = useState(lesson.tag || 'blue');
   const updateLesson = useStore(({ lessonSlice }) => lessonSlice.updateLesson);
   const deleteLesson = useStore(({ lessonSlice }) => lessonSlice.deleteLesson);
+  const fetchAllLessons = useStore(({ lessonSlice }) => lessonSlice.fetchAllLessons);
+  const isAuth = useStore(({ authSlice }) => authSlice.authenticated);
   const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
   const [isColorModalOpen, setIsColorModalOpen] = useState(false);
   const [isAddToFolderModalOpen, setIsAddToFolderModalOpen] = useState(false);
   const [isRemoveFromFolderModalOpen, setIsRemoveFromFolderModalOpen] = useState(false);
   const [folderSearch, setFolderSearch] = useState('');
   const [selectedFolders, setSelectedFolders] = useState([]);
+  const [isRemixModalOpen, setIsRemixModalOpen] = useState(false);
   const user = useStore(({ userSlice }) => userSlice.current);
   const folders = user?.folders || {};
   const addLessonToFolder = useStore(({ userSlice }) => userSlice.addLessonToFolder);
   const deleteLessonFromFolder = useStore(({ userSlice }) => userSlice.deleteLessonFromFolder);
+  const createLesson = useStore(({ lessonSlice }) => lessonSlice.createLesson);
+  const fetchLesson = useStore(({ lessonSlice }) => lessonSlice.fetchLesson);
   const { colors } = useTheme();
-
   // Filter folders by search
   const filteredFolders = Object.keys(folders).filter((folder) => folder.toLowerCase().includes(folderSearch.toLowerCase()));
 
@@ -50,31 +54,67 @@ function LessonCard({ lesson, onDelete }) {
   const availableFolders = filteredFolders.filter((folder) => !lessonFolders.includes(folder));
 
   const handleAddToFolder = async () => {
+    if (selectedFolders.length === 0) {
+      toast({
+        title: 'No Folders Selected',
+        description: 'Please select at least one folder to add the lesson to',
+        status: 'warning',
+        duration: 3000,
+        isClosable: true,
+      });
+      return;
+    }
     try {
       await Promise.all(selectedFolders.map((folder) => addLessonToFolder(folder, lesson._id || lesson.id)));
       toast({
-        title: 'Lesson added to folders', status: 'success', duration: 2000, isClosable: true,
+        title: 'Lesson added to folders',
+        description: `Added to ${selectedFolders.length} folder${selectedFolders.length > 1 ? 's' : ''}`,
+        status: 'success',
+        duration: 2000,
+        isClosable: true,
       });
       setIsAddToFolderModalOpen(false);
       setSelectedFolders([]);
     } catch (error) {
       toast({
-        title: 'Error', description: error.message, status: 'error', duration: 3000, isClosable: true,
+        title: 'Error',
+        description: error.message,
+        status: 'error',
+        duration: 3000,
+        isClosable: true,
       });
     }
   };
 
   const handleRemoveFromFolder = async () => {
+    if (selectedFolders.length === 0) {
+      toast({
+        title: 'No Folders Selected',
+        description: 'Please select at least one folder to remove the lesson from',
+        status: 'warning',
+        duration: 3000,
+        isClosable: true,
+      });
+      return;
+    }
     try {
       await Promise.all(selectedFolders.map((folder) => deleteLessonFromFolder(folder, lesson._id || lesson.id)));
       toast({
-        title: 'Lesson removed from folders', status: 'success', duration: 2000, isClosable: true,
+        title: 'Lesson removed from folders',
+        description: `Removed from ${selectedFolders.length} folder${selectedFolders.length > 1 ? 's' : ''}`,
+        status: 'success',
+        duration: 2000,
+        isClosable: true,
       });
       setIsRemoveFromFolderModalOpen(false);
       setSelectedFolders([]);
     } catch (error) {
       toast({
-        title: 'Error', description: error.message, status: 'error', duration: 3000, isClosable: true,
+        title: 'Error',
+        description: error.message,
+        status: 'error',
+        duration: 3000,
+        isClosable: true,
       });
     }
   };
@@ -83,6 +123,28 @@ function LessonCard({ lesson, onDelete }) {
     setSelectedFolders((prev) => (prev.includes(folder)
       ? prev.filter((f) => f !== folder)
       : [...prev, folder]));
+  };
+
+  const handleCreateCopy = async () => {
+    try {
+      const lessonCopy = await createLesson({
+        title: lesson.title,
+        objectives: lesson.objectives,
+        overview: lesson.overview,
+        materials: lesson.materials,
+        steps: lesson.steps,
+        standards: lesson.standards,
+        grade: lesson.grade,
+        subject: lesson.subject,
+        status: 'public',
+        shared: [lesson.author.id],
+        tag: lesson.tag,
+      });
+      return lessonCopy;
+    } catch (error) {
+      console.error('Error remixing lesson');
+      return false;
+    }
   };
 
   // added these fns bc both color and delete modals were opening simultaneously
@@ -116,9 +178,7 @@ function LessonCard({ lesson, onDelete }) {
         tag: color,
       });
       setSelectedColor(color);
-      // Close modal after selected
       closeColorModal();
-      // Show success
       toast({
         title: 'Color Updated',
         description: `Lesson color changed to ${color}`,
@@ -126,10 +186,10 @@ function LessonCard({ lesson, onDelete }) {
         duration: 2000,
         isClosable: true,
       });
+      await fetchAllLessons(isAuth);
     } catch (error) {
       console.error('Error selecting color:', error);
 
-      // Show error
       toast({
         title: 'Color Update Failed',
         description: 'Unable to update lesson color',
@@ -140,18 +200,52 @@ function LessonCard({ lesson, onDelete }) {
     }
   };
 
-  const handleView = (event) => {
+  const handleView = async (event) => {
     event.preventDefault();
-    navigate(`/edit/${lesson._id}?tab=1`);
+    try {
+      await fetchLesson(lesson._id);
+      navigate(`/view/${lesson._id}`);
+    } catch (error) {
+      console.error('Error loading lesson:', error);
+    }
   };
 
-  const handleEdit = (event) => {
+  const handleEdit = async (event) => {
     event.preventDefault();
-    navigate(`/edit/${lesson._id}?tab=0`);
+    if (!isAuth) {
+      toast({
+        title: 'Please log in',
+        description: 'Please log in or sign up to make changes.',
+        duration: 3000,
+        isClosable: true,
+      });
+      return;
+    }
+    // If we try to edit a lesson that isn't ours
+    if (user?.username !== lesson?.author?.username) {
+      setIsRemixModalOpen(true);
+    } else {
+      try {
+        await fetchLesson(lesson._id);
+        navigate(`/edit/${lesson._id}?tab=1`);
+      } catch (error) {
+        console.error('Error loading lesson:', error);
+      }
+    }
   };
 
   const handleDelete = async () => {
     try {
+      if (user?.id !== lesson?.author?.id) {
+        toast({
+          title: 'Invalid Delete',
+          description: 'Cannot delete a lesson not created by you',
+          status: 'error',
+          duration: 2000,
+          isClosable: true,
+        });
+        return;
+      }
       await deleteLesson(lesson.id);
       toast({
         title: 'Lesson Deleted',
@@ -192,7 +286,6 @@ function LessonCard({ lesson, onDelete }) {
       bg={colors.cardBg}
       boxShadow="md"
       borderLeftWidth="6px"
-      borderLeftColor={`${selectedColor}.500`}
       _hover={{ boxShadow: 'lg', transform: 'translateY(-2px)' }}
       transition="all 0.2s"
       position="relative"
@@ -202,25 +295,24 @@ function LessonCard({ lesson, onDelete }) {
       _active={{ cursor: 'grabbing' }}
       display="flex"
       flexDirection="column"
+      onClick={handleView}
     >
 
       {/* Lesson image & title */}
       <Box
         height="100px"
-        backgroundImage='linear-gradient(rgba(62, 65, 69, 0.59)), url("https://picsum.photos/320/180")'
-        backgroundSize="cover"
-        backgroundPosition="center"
-        backgroundRepeat="no-repeat"
+        bgGradient={`linear(135deg, ${selectedColor}.500), ${selectedColor}.600`}
         display="flex"
         alignItems="center"
         justifyContent="center"
+        borderTopRadius="md"
       >
         <Text
           fontWeight="bold"
           fontSize="xl"
           color="white"
           textAlign="center"
-          textShadow="2px 2px 4px rgba(0,0,0,0.8)"
+          textShadow="2px 2px 4px rgba(0,0,0,0.5)"
           px={4}
           maxW="220px"
           noOfLines={2}
@@ -243,7 +335,7 @@ function LessonCard({ lesson, onDelete }) {
         </Text>
       </Box>
 
-      {/** Author Avatar & Name */}
+      {/* Author Avatar & Name */}
 
       <Box
         position="absolute"
@@ -257,15 +349,14 @@ function LessonCard({ lesson, onDelete }) {
       >
         <Avatar
           size="sm"
-          name={lesson.author.username}
-          src="https://picsum.photos/40/40"
+          name={lesson?.author?.username}
         />
         <Text fontSize="sm" color={colors.text} fontWeight="medium">
-          {lesson.author.username}
+          {lesson?.author?.username}
         </Text>
       </Box>
 
-      {/** Menu */}
+      {/* Menu */}
 
       <Menu>
         <MenuButton
@@ -278,24 +369,67 @@ function LessonCard({ lesson, onDelete }) {
           right="2"
           aria-label="Options"
           color={colors.text}
+          onClick={(e) => e.stopPropagation()}
         />
         <Portal>
-          <MenuList zIndex="10" position="relative" bg={colors.modalBg}>
-            <MenuItem onClick={handleView} bg={colors.modalBg} color={colors.text} _hover={{ bg: colors.hover }}>View</MenuItem>
-            <MenuItem onClick={handleEdit} bg={colors.modalBg} color={colors.text} _hover={{ bg: colors.hover }}>Edit</MenuItem>
-            <MenuItem onClick={openColorModal} bg={colors.modalBg} color={colors.text} _hover={{ bg: colors.hover }}>
+          <MenuList zIndex="10" position="relative" bg={colors.modalBg} onClick={(e) => e.stopPropagation()}>
+            <MenuItem onClick={(e) => { e.stopPropagation(); handleView(e); }} bg={colors.modalBg} color={colors.text} _hover={{ bg: colors.hover }}>View</MenuItem>
+            <MenuItem onClick={(e) => { e.stopPropagation(); handleEdit(e); }} bg={colors.modalBg} color={colors.text} _hover={{ bg: colors.hover }}>Edit</MenuItem>
+            <MenuItem onClick={(e) => { e.stopPropagation(); openColorModal(); }} bg={colors.modalBg} color={colors.text} _hover={{ bg: colors.hover }}>
               Change Lesson Color
             </MenuItem>
-            <MenuItem onClick={() => setIsAddToFolderModalOpen(true)} bg={colors.modalBg} color={colors.text} _hover={{ bg: colors.hover }}>
+            <MenuItem onClick={(e) => { e.stopPropagation(); setIsAddToFolderModalOpen(true); }} bg={colors.modalBg} color={colors.text} _hover={{ bg: colors.hover }}>
               Add to Folder
             </MenuItem>
-            <MenuItem onClick={() => setIsRemoveFromFolderModalOpen(true)} bg={colors.modalBg} color={colors.text} _hover={{ bg: colors.hover }}>
+            <MenuItem onClick={(e) => { e.stopPropagation(); setIsRemoveFromFolderModalOpen(true); }} bg={colors.modalBg} color={colors.text} _hover={{ bg: colors.hover }}>
               Remove from Folder
             </MenuItem>
-            <MenuItem onClick={openDeleteModal} bg={colors.modalBg} color="red.500" _hover={{ bg: colors.hover }}>
+            <MenuItem onClick={(e) => { e.stopPropagation(); openDeleteModal(); }} bg={colors.modalBg} color="red.500" _hover={{ bg: colors.hover }}>
               Delete
             </MenuItem>
 
+            <Modal
+              isOpen={isRemixModalOpen}
+              onClose={() => {
+                setIsRemixModalOpen(false);
+              }}
+              size="md"
+            >
+              <ModalOverlay />
+              <ModalContent bg={colors.modalBg}>
+                <ModalHeader color={colors.text}>Remix Lesson</ModalHeader>
+                <ModalCloseButton color={colors.text} />
+                <ModalBody color={colors.text}>
+                  <Text mb={3}>
+                    {'You are about to remix someone else\'s lesson. This will create a copy that you can edit. The original author will be able to view your new lesson.'}
+                  </Text>
+                </ModalBody>
+                <ModalFooter>
+                  <Button
+                    variant="ghost"
+                    onClick={() => {
+                      setIsRemixModalOpen(false);
+                    }}
+                    mr={3}
+                    color={colors.text}
+                  >
+                    Cancel
+                  </Button>
+                  <Button
+                    colorScheme="blue"
+                    onClick={async () => {
+                      const lessonCopy = await handleCreateCopy();
+                      if (lessonCopy) {
+                        navigate(`/edit/${lessonCopy._id}?tab=1`);
+                      }
+                      setIsRemixModalOpen(false);
+                    }}
+                  >
+                    Continue
+                  </Button>
+                </ModalFooter>
+              </ModalContent>
+            </Modal>
             <Modal
               isOpen={isColorModalOpen}
               onClose={closeColorModal}
@@ -314,7 +448,7 @@ function LessonCard({ lesson, onDelete }) {
                         height="50px"
                         borderRadius="md"
                         cursor="pointer"
-                        onClick={() => selectColor(color)}
+                        onClick={(e) => { e.stopPropagation(); selectColor(color); }}
                         _hover={{
                           transform: 'scale(1.1)',
                           boxShadow: 'lg',
@@ -379,7 +513,7 @@ function LessonCard({ lesson, onDelete }) {
                     {availableFolders.map((folder) => (
                       <Button
                         key={folder}
-                        onClick={() => toggleFolderSelection(folder)}
+                        onClick={(e) => { e.stopPropagation(); toggleFolderSelection(folder); }}
                         bg={selectedFolders.includes(folder) ? 'blue.300' : 'blue.100'}
                         _hover={{ bg: 'blue.300' }}
                         color={colors.text}
@@ -410,7 +544,7 @@ function LessonCard({ lesson, onDelete }) {
                     {lessonFolders.map((folder) => (
                       <Button
                         key={folder}
-                        onClick={() => toggleFolderSelection(folder)}
+                        onClick={(e) => { e.stopPropagation(); toggleFolderSelection(folder); }}
                         bg={selectedFolders.includes(folder) ? 'blue.300' : 'blue.100'}
                         _hover={{ bg: 'blue.300' }}
                         color={colors.text}
